+++ ChangeLog.md
@@ -3,6 +3,8 @@ GeoIP database ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Fixed issue #1: Error message when input is not an IP address - @thekid
+
 ## 0.1.0 / 2015-10-02
 
 * First public release - @thekid